Minor enhancements
Minor changes to the register time. Both front end and back end.
I have optimized the timeInterval to be a 2 dimensional array. So when the user selects the time, in the registration a tmpTime variable holds the time in minutes.
I have also changes front end to display succes or error. And the service responds with the correct form of reponse.