Create response class for php services
I have created the response class and I have changed in the calls so it handles the data almost the same way. It depends on if you put responseType as json or text. If you put it as text you can use JSON.parse(response) directly on reponse.
If you use json you have to do it like this: JSON.parse(JSON.stringify(response)). This way you can can handle the object with dot notation like: resp.success == true
or
this.user = resp.result;