#29 - Register Time - Filter borger list
I have made the top list item to be an input field in the borger list drop down menu.
I started making it in the dropdown button it self, which is possible, but it gives some focusing issues.
When the user epands the borger list, the first item is an input field, when the user starts typing, the borger list will automatically update.