#30 - Register Time - Date selection
1 Hour
Started on changing the date selection in register time. With this new module the user selects date and time for start and end.
Right now they are hidden, because the database needs to be changed to hold this information as well as the old. We don't want to delete the other way completely, just hide it for now. We are not sure what will be best.