#19 - Home page
I got the filter on the recent activity table to work. So you can search as you can on other tables, but you can also choose a period where you want to see registrations from.

I also got the download button to work. So when you click on the excel button next to the settings button, you will download an excel sheet with the data shown in the table.
(I am having issues where it says it is untrusted when I download it, I dont know why or if I'm the only one.)