#38 - Fix cookie token
Cookie token is only generated and saved in the database the first time the user logs in. In the future it will use the already stored token.

If the user already has a token in the database, it uses this on as the token on the device. If token on the other hand is empty, it will generate one check if it exist, if not it will save it.